tw_bf behavior - reset after rc calls
In optimistic mode, an event can be rolled back and reapplied multiple times.
Currently, the bitfield on rollbacks after the first contain bits set from all
previous events and is not reset by ROSS in between. This change uses the more
intuitive behavior of a tw_bf's state only persisting for a single
event-rollback sequence, by resetting the bitfield at the end of the reverse
event function call.

Note that I did not touch the ROSS models, some of which manually clear out the
bitfields or use it in strange ways (pm-node.c casts to int and does
arithmetic???). I'll leave those for the model maintainers.