Create core RNG array for exclusive ROSS engine usage
This commit adds a separate array of RNG streams on each LP that aren't
to be utilized by developed models. These separate RNG streams can be
utilized to leverage the deterministic RNG nature that ROSS can manage
toward other goals of the ROSS engine itself.

Notable example use for this: Deterministic Tiebreaking
Deterministic Tiebreaking can be implemented by creating a random value
at the creation of an event, this value is encoded into the ROSS event
struct and is utilized to break any event ties (same destination LP at
same time). Because this separate RNG is only accessed by ROSS, it can
be rolled back if the event becomes RC'd or cancelled. Because of
determinism, any ordering as a result of this tiebreaker will be
consistent across simulation runs regardless of event delivery order
or stragglers. If a regular model-accessed LP RNG was used for this
purpose, the tiebreaking sequence would be subject to interference.