Implement new time signature paradigm and deterministic tiebreaker
This commit adds the functionality of the deterministic tiebreaker
mentioned in an earlier commit which added the core ROSS engine
exclusive RNGs.

The deterministic tiebreaker itself is rather simple. When an event
is created, a random value is generated from a ROSS core RNG stream.
When that event is RC'd or cancelled, that random value is also
reversed. Because this stream is only utilized by said tiebreaking
mechanism, the ordering of tiebreaking values created by the stream
is deterministic across simulations. When comparing two events
received by an LP at the same timestamp, the determining factor
in which is processed first will be decided - deterministically -
by the tiebreaker value.

While the concept itself is simple, and implementing the tiebreaker
into the event struct is similarly simple, getting this tiebreaker
to work with the concept of GVT and rollbacks is not.

A better way to think about this tiebreaking mechanism is to think
of it as making sure that there are actually no such thing as event
ties. This paradigm shift means that determining "when" GVT happens
is no longer a single TW_STIME value. There is now an event signature
struct which contains a timestamp and a tiebreaker value. This
signature is all that is necessary for determining ordering of two
events in the simulation. Thus the time of the last GVT is no longer
just the single dimensional virtual timestamp, but it also includes
a tiebreaker which divides events that happen at the same primary
timestamp as GVT but with their own tiebreaker values which will be
deterministically separated as "before GVT" and "after GVT".

Thus, rollbacks now also no longer go back to a single timestamp
value in time, but to a two dimensional timestamp value consisting
of the primary timestamp and an event tiebreaker.

As complex as this system is, it does have its benefits:
1) Comparing events for Splay and AVL trees to determine ordering
require fewer comparisons and thus less compute time spent.
2) If primary timestamp ties are numerous in a model, rolling back
of one event at said timestamp will no longer require rolling back
all events with the same timestamp, only those whose tiebreaker values
determine that they happen "after" the event that prompted the
rollback.
3) Event ties are statistically impossible to force. Because the
tiebreaker value is generated using its own independent RNG stream
with an extremely long period, two events at the same primary
timestamp ALSO generating an identical random value is nearly
impossible. This also means that model developers will no longer
have to generate their own small noise to add onto their event
timestamps to prevent event ties, significantly improving the
administrative code complexity - reducing the likelihood that
a developer will forget to roll back an RNG from noise and
plunge their entire model into non-determinism.

This feature has been walled off behind a CMAKE Define Variable:
USE_RAND_TIEBREAKER. Set this value to ON during CMAKE ROSS
configuration and all code enabling the tiebreaking value
generation and the timestamp-to-time-signature paradigm shift
will be switched on by pre-processor #ifdef's.