Implementing new synchronizaton algorithm: sequential rollback check
With this new mode, one can test if all event reverse handlers have
been properly implemented. It will run each event twice, rollbacking it
once, and it will check that:

(1) processing and rollbacking an event will produce the same LP state
  as not processing the event at all, and
(2) processing the event after rollbacking it is the same as just
  processing the event once

Additionally to checking the LP state, we also check the state of random
generators.

The runtime penalty of processing all events twice is an increase of at
least 100%. In phold experiments, this mode takes about 2.8x times the
original.