Implementing proper GVT operation for tie-breaker mechanism
Previously, regardless of whether the tie-breaker mechanism was on or
off, we would compute GVT using `recv_ts` only (a double). This will
rarely be an issue, but it is incorrect. We expect GVT to be the
smallest timestamp that hasn't been processed yet, but in the case of
the tie-breaker mechanism, it would never correspond to any event in the
whole simulation.

The solution is to use the whole `struct tw_event_sig` in the GVT
computation. For this, we needed to implement the minimum operation for
tw_event_sig in MPI.